6-size.c

+++ 0x00-hello_world/6-size.c
@@ -8,10 +8,10 @@
 
 int main(void)
 {
-	printf("Size of a char: %d bytes(s)", (unsigned long)sizeof(char));
-	printf("Size of an int: %d bytes(s)", (unsigned long)sizeof(int));
-	printf("Size of a long int: %d bytes(s)", (unsigned long)sizeof(long int));
-	printf("Size of a long long int: %d bytes(s)", (unsigned long)sizeof(long long int));
-	printf("Size of a float: %d bytes(S)", (unsigned long)sizeof(float));
+	printf("Size of a char: %d bytes(s)", sizeof(char));
+	printf("Size of an int: %d bytes(s)", sizeof(int));
+	printf("Size of a long int: %d bytes(s)", sizeof(long int));
+	printf("Size of a long long int: %d bytes(s)", sizeof(long long int));
+	printf("Size of a float: %d bytes(S)", sizeof(float));
 	return (0);
 }
